convert chainwork to gigahashes

+++ database/migrations/2020_05_25_200120_create_blocks_table.php
@@ -35,7 +35,7 @@ class CreateBlocksTable extends Migration
             $table->string('bits');
             $table->double('difficulty');
             $table->double('hashrate')->nullable();
-            $table->decimal('chainwork', 20, 0);
+            $table->unsignedBigInteger('chainwork');
             $table->string('previousblockhash')->nullable();
 
             $table->timestamps();
